+++ app/src/main/java/com/soumikshah/investmenttracker/view/GraphFragment.java
@@ -1,0 +1,28 @@
+package com.soumikshah.investmenttracker.view;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
+import com.soumikshah.investmenttracker.R;
+
+public class GraphFragment extends Fragment {
+
+    public GraphFragment(){}
+
+    @Override
+    public void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.graphfragment, container, false);
+    }
+}
