+++ app/src/main/res/layout/mainfragment.xml
@@ -1,0 +1,9 @@
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context="com.soumikshah.investmenttracker.view.MainFragment">
+
+    <include layout="@layout/content_main" />
+
+</RelativeLayout>
